better store use (FINALLY)
